refactor: split helpers of driver into a new object library
* To use in lpac frontend code.
* To add lpac path prefix to avoid common header name conflict.
