+++ CMakeLists.txt
@@ -67,5 +67,6 @@ endif()
 
 add_subdirectory(cjson)
 add_subdirectory(euicc)
+add_subdirectory(utils)
 add_subdirectory(driver)
 add_subdirectory(src)
