+++ driver/CMakeLists.txt
@@ -18,10 +18,9 @@ if(LPAC_DYNAMIC_DRIVERS)
 else()
     add_library(euicc-drivers STATIC ${DIR_INTERFACE_SRCS})
 endif()
-target_link_libraries(euicc-drivers euicc cjson-static)
+target_link_libraries(euicc-drivers euicc cjson-static lpac-utils)
 target_include_directories(euicc-drivers PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
-target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/helpers.c)
 target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/apdu/stdio.c)
 target_sources(euicc-drivers PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/http/stdio.c)
