+++ driver/apdu/at_common.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <helpers.h>
+#include <lpac/utils.h>
 
 #define AT_BUFFER_SIZE 20480
 #define AT_READ_BUFFER_SIZE 4096
