+++ utils/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_library(lpac-utils OBJECT lpac/utils.c)
+target_include_directories(lpac-utils PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_link_libraries(lpac-utils PRIVATE cjson-static)
