+++ utils/lpac/utils.h
